wip fix PresentationApiTests

+++ wrappers/dotnet/anoncreds-rs-dotnet/Models/Structures.cs
@@ -225,8 +225,8 @@ namespace anoncreds_rs_dotnet.Models
         public struct FfiNonrevokedIntervalOverride
         {
             public FfiStr RevRegDefId;
-            public long RequestedFromTs;
-            public long OverrideRevStatusListTs;
+            public int RequestedFromTs;
+            public int OverrideRevStatusListTs;
 
             public static FfiNonrevokedIntervalOverride Create(NonrevokedIntervalOverride intervalOverride)
             {
